added if for staging/master

+++ Jenkinsfile
@@ -20,12 +20,19 @@ pipeline {
             }
         }
         stage('Deploy') {
-            steps {
-                sh 'echo "Deploy"'
-                sh 'docker kill Petclinic ||true'
-                sh 'docker rm Petclinic ||true'
-                sh 'docker run -d -i -t -p 8080:8080 --name Petclinic petclinic'
-            }
+              if (env.BRANCH_NAME == "master") {                                          
+                    steps {
+                    sh 'echo "Deploy"'
+                    sh 'docker kill Petclinic ||true'
+                    sh 'docker rm Petclinic ||true'
+                    sh 'docker run -d -i -t -p 8080:8080 --name Petclinic petclinic'
+                    }
+            } else {                                   
+                    steps {
+                    sh 'echo "Artifact was built"'
+                    }
+            } 
+            
         }
     }
 }
